Refactor nav settings popover
